+++ src/global.d.ts
@@ -3,6 +3,10 @@
 // Definitions by: [Alex Regan] <[https://github.com/alexsasharegan/]>
 
 declare module "*.vue" {
-  import Vue from "vue"
-  export default Vue
+  import Vue from "vue";
+  export default Vue;
 }
+
+declare type AnyObject = { [key: string]: any };
+
+declare type Dictionary<T> = { [key: string]: T };