+++ test/component.test.ts
@@ -1,0 +1,5 @@
+describe("Vue Transmit", () => {
+	it("should work", async () => {
+		// 🔬
+	})
+})