+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="9.0.0-alpha.1"></a>
+# [9.0.0-alpha.1](https://github.com/alexsasharegan/vue-transmit/compare/v9.0.0-alpha.0...v9.0.0-alpha.1) (2018-02-09)
+
+
+### Bug Fixes
+
+* correct constructor prop for adapter ([e1807fc](https://github.com/alexsasharegan/vue-transmit/commit/e1807fc))
+
+
+
+<a name="9.0.0-alpha.0"></a>
+# [9.0.0-alpha.0](https://github.com/alexsasharegan/vue-transmit/compare/v8.0.0...v9.0.0-alpha.0) (2018-02-09)
+
+
+### Bug Fixes
+
+* removing vue-class-component for smaller bundles ([9bf830c](https://github.com/alexsasharegan/vue-transmit/commit/9bf830c))
+
+
+### Code Refactoring
+
+* **classes:** use named export for VTransmitFile ([78004dc](https://github.com/alexsasharegan/vue-transmit/commit/78004dc))
+
+
+### Features
+
+* **build:** use rollup for smaller build ([8f64477](https://github.com/alexsasharegan/vue-transmit/commit/8f64477))
+* **xhr-transport:** initial adapter complete ([93e16a8](https://github.com/alexsasharegan/vue-transmit/commit/93e16a8))
+
+
+### BREAKING CHANGES
+
+* **build:** browser bundle & commonjs bundle
+* **classes:** switch to named export of VTransmitFile class
+
+
+
 <a name="8.0.0"></a>
 # [8.0.0](https://github.com/alexsasharegan/vue-transmit/compare/v7.1.7...v8.0.0) (2018-01-12)